Refactor tests and add more tests for addMarkdownSyntax helper
